requirements.txt
Django==4.0.1
django-environ==0.8.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
mango-jwt==1.3.2
mongoengine==0.23.1
PyJWT==2.3.0
pymongo==3.12.3